printf

+++ format/ft_printf_format_c.c
@@ -6,13 +6,13 @@
 /*   By: kipark <kipark@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/21 19:05:13 by kipark            #+#    #+#             */
-/*   Updated: 2022/02/27 18:13:05 by kipark           ###   ########seoul.kr  */
+/*   Updated: 2022/02/27 21:39:41 by kipark           ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include"ft_printf_format.h"
+#include"ft_printf.h"
 
-ssize_t	ft_printf_format_c(va_list ap)
+int	ft_printf_format_c(va_list ap)
 {
 	char	ap_result;
